Created different files for migrating different parts of AidStream
